Refactor: restructure package and type naming conventions
Renamed packages and files for better organization and consistency. Simplified type names like `componentsPool` to `Pool` and `CommandManager` to `Manager` to align with standard naming practices. This improves maintainability and readability across the codebase.
