--- app/backend/command/CommandManager.go
@@ -1,4 +1,0 @@
-package command
-
-type CommandManager struct {
-}
