+++ app/backend/command/manager.go
@@ -1,0 +1,4 @@
+package command
+
+type Manager struct {
+}
