+++ app/backend/components/pool.go
@@ -4,15 +4,15 @@ import (
 	"Dr.uml/backend/Component"
 )
 
-type componentsPool struct {
+type Pool struct {
 	list []*component.Component
 }
 
-func (cp *componentsPool) insert(c *component.Component) {
+func (cp *Pool) insert(c *component.Component) {
 	cp.list = append(cp.list, c)
 }
 
-func (cp *componentsPool) remove(c *component.Component) {
+func (cp *Pool) remove(c *component.Component) {
 	for i, comp := range cp.list {
 		if comp == c {
 			cp.list = append(cp.list[:i], cp.list[i+1:]...)
@@ -21,6 +21,6 @@ func (cp *componentsPool) remove(c *component.Component) {
 	}
 }
 
-func (cp *componentsPool) Len() int {
+func (cp *Pool) Len() int {
 	return len(cp.list)
 }
