+++ app/backend/utils/error/invalidArgument.go
@@ -1,4 +1,4 @@
-package utils
+package error
 
 type InvalidArgumentError struct {
 	msg string
