+++ app/backend/utils/error/send.go
@@ -1,4 +1,4 @@
-package utils
+package error
 
 type SendError struct {
 	msg string
